handle prologue view in different liveview + redirect to lesson view when beginning an attempt
