+++ lib/oli_web/router.ex
@@ -979,7 +979,7 @@ defmodule OliWeb.Router do
         live("/practice", Delivery.Student.PracticeLive)
       end
 
-      scope "/lesson/:revision_slug" do
+      scope "/prologue/:revision_slug" do
         live_session :delivery_lesson,
           root_layout: {OliWeb.LayoutView, :delivery},
           layout: {OliWeb.Layouts, :student_delivery_lesson},
@@ -987,6 +987,24 @@ defmodule OliWeb.Router do
             OliWeb.LiveSessionPlugs.SetUser,
             OliWeb.LiveSessionPlugs.SetSection,
             {OliWeb.LiveSessionPlugs.InitPage, :set_page_context},
+            OliWeb.LiveSessionPlugs.SetBrand,
+            OliWeb.LiveSessionPlugs.SetPreviewMode,
+            OliWeb.LiveSessionPlugs.RequireEnrollment,
+            OliWeb.LiveSessionPlugs.SetRequestPath
+          ] do
+          live("/", Delivery.Student.PrologueLive)
+        end
+      end
+
+      scope "/lesson/:revision_slug" do
+        live_session :delivery_prologue,
+          root_layout: {OliWeb.LayoutView, :delivery},
+          layout: {OliWeb.Layouts, :student_delivery_lesson},
+          on_mount: [
+            OliWeb.LiveSessionPlugs.SetUser,
+            OliWeb.LiveSessionPlugs.SetSection,
+            {OliWeb.LiveSessionPlugs.InitPage, :set_page_context},
+            OliWeb.LiveSessionPlugs.RedirectToPrologue,
             OliWeb.LiveSessionPlugs.RedirectAdaptiveChromeless,
             OliWeb.LiveSessionPlugs.SetBrand,
             OliWeb.LiveSessionPlugs.SetPreviewMode,
